requirements.txt
flask==2.2.3
werkzeug==2.2.3
beautifulsoup4==4.12.2
requests==2.28.2
python-dotenv==1.0.0
gunicorn==20.1.0
gradio_client==0.7.0
lxml==4.9.3 